Add utilization reporting
Returns a list of utilization rates per hour for a target node/computer
on a Jenkins host. When a discontinuity is detected in the node (zero
executors available) the history is halted. The list containes rates in
descending order by date/time and represents average utilization rates
in decimal form as a fraction of 1.0.